backend/requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pyserial==3.4
PyYAML==3.12
six==1.11.0
Werkzeug==0.13
